Rename the Project and source files from git-changelog to git-cl
This just renames everthing from git-changelog to git-cl

[chaneglog]
changed: Renamed the project from git-changelog to git-cl. This was part of the repo change name.
changed: Cleaned up everything and removed a bunch of code that wasn't used.

ps-id: 5EA1D48F-C1BA-4465-B679-17DF8FB10D12
