+++ Sources/git-cl/Bash.swift
