+++ Sources/git-cl/RunResult.swift
