+++ Tests/GitCLTests/XCTestManifests.swift
@@ -3,7 +3,7 @@ import XCTest
 #if !canImport(ObjectiveC)
 public func allTests() -> [XCTestCaseEntry] {
     return [
-        testCase(git_changelogTests.allTests),
+        testCase(GitCLTests.allTests),
     ]
 }
 #endif
