+++ Tests/LinuxMain.swift
@@ -3,5 +3,5 @@ import XCTest
 import git_changelogTests
 
 var tests = [XCTestCaseEntry]()
-tests += git_changelogTests.allTests()
+tests += GitCLTests.allTests()
 XCTMain(tests)
